♻️ refactor : App.js 파일 정리
