+++ src/utils/calcPurchaseQuantity.js
@@ -1,4 +1,4 @@
-import { PURCHASE_AMOUNT_UNIT } from "../utils/constants.js";
+import { PURCHASE_AMOUNT_UNIT } from "./constants.js";
 
 export const calcPurchaseQuantity = (input) => {
   const PURCHASE_QUANTITY = input / PURCHASE_AMOUNT_UNIT;
